+++ common/src/main/java/pepjebs/mapatlases/client/MapAtlasesClient.java
@@ -59,13 +59,15 @@ public class MapAtlasesClient {
 
     public static final Material MAP_HOVERED_TEXTURE = new Material(
             new ResourceLocation("textures/atlas/shulker_boxes.png"), //so we have mipmap here too
-            MapAtlasesMod.res("gui/screen/map_border"));
+            MapAtlasesMod.res("gui/screen/map_hovered"));
 
     public static final ResourceLocation MAP_ICON_TEXTURE = new ResourceLocation("textures/map/map_icons.png");
     public static final ResourceLocation ATLAS_OVERLAY_TEXTURE = MapAtlasesMod.res("textures/gui/screen/atlas_overlay.png");
     public static final ResourceLocation ATLAS_BACKGROUND_TEXTURE = MapAtlasesMod.res("textures/gui/screen/atlas_background.png");
     public static final ResourceLocation ATLAS_BACKGROUND_TEXTURE_BIG = MapAtlasesMod.res("textures/gui/screen/atlas_background_big.png");
     public static final ResourceLocation GUI_ICONS_TEXTURE = new ResourceLocation("textures/gui/icons.png");
+    public static final ResourceLocation MAP_HUD_BACKGROUND_TEXTURE = MapAtlasesMod.res("textures/gui/hud/map_background.png");
+    public static final ResourceLocation MAP_HUD_FOREGROUND_TEXTURE = MapAtlasesMod.res("textures/gui/hud/map_foreground.png");
 
 
     public static final List<String> DIMENSION_TEXTURE_ORDER = List.of(Level.OVERWORLD.location().toString(),
@@ -137,7 +139,7 @@ public class MapAtlasesClient {
             maps.addNotSynced(player.level());
             Slice slice = MapAtlasItem.getSelectedSlice(atlas, player.level().dimension());
             // I hate this
-            currentActiveMapKey = MapKey.at(maps.getScale(), player, slice);
+            currentActiveMapKey = MapKey.containing(maps.getScale(), player, slice);
             MapDataHolder select = maps.select(currentActiveMapKey);
             if (select == null) {
                 select = maps.getClosest(player, slice);